Added style for calendar, and made screen navigation smoother
The smoother navigation is, in fact, just resetting the views once opening a new screen.
One example is the calendar in the task maker, before if the user opened the calendar and navigated to a new screen, the calendar would not be closed. Now that has been fixed and behaves as expected.