Fixed a few bugs found on the emulator
Fixed a few bugs that I have encountered while recording the video on an Android emulator.
The fixes where fast, I am not sure if I made more bugs by trying to fix the ones I had...
But at this point the app is stable enough to work on a device, with no crashes or errors.